Add BSP directory and move files
