+++ andino_firmware/src/bsp/interrupt_in_arduino.cpp
@@ -27,7 +27,7 @@
 // CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-#include "interrupt_in_arduino.h"
+#include "andino/bsp/interrupt_in_arduino.h"
 
 #include <Arduino.h>
